% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_correlations.R
\name{calculate_correlations}
\alias{calculate_correlations}
\title{Calculate correlation between two data frames}
\usage{
calculate_correlations(
  df1,
  df2,
  method = "pearson",
  adjust_method = "fdr",
  use = "all.obs",
  show_significance = "stars"
)
}
\arguments{
\item{df1}{A data frame or matrix containing the first set of variables
to be correlated. Rows are samples, columns are variables
(e.g., genes, proteins, metabolites).}

\item{df2}{A data frame or matrix containing the second set of variables
to be correlated. Rows are samples, columns are variables.}

\item{method}{The correlation method to be used. Default is "pearson".
Other options include "spearman" and "kendall".}

\item{adjust_method}{Method for adjusting the p-values for multiple
testing. Default is "fdr". If "none", no adjustment is applied.
Supports all methods from \code{p.adjust.methods}.}

\item{use}{Handling of missing data. One of "all.obs" or
"pairwise.complete.obs".}

\item{show_significance}{How to represent significance in the output.
One of "stars", "p_value", or "correlation".}
}
\value{
A list containing:
\item{correlation}{Matrix of correlation coefficients (rounded to 3 decimals).}
\item{p_value}{Matrix of raw p-values.}
\item{p_value_adj}{Matrix of adjusted p-values.}
\item{signif_matrix}{Matrix of significance stars, p-values, or correlations.}
}
\description{
This function calculates the correlation matrix between the columns
of two data frames.
}
\details{
Pearson and Spearman correlations are computed using
\code{WGCNA::corAndPvalue} for efficiency.
Kendall correlation uses \code{cor.test} in a loop and may be slow
for large datasets.
}
\examples{
data(Transcriptomics)
data(Metagenomics)

result_list <- calculate_correlations(
  df1 = Transcriptomics,
  df2 = Metagenomics,
  show_significance = "p_value"
)

}
