% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmniCorr-package.R
\docType{package}
\name{OmniCorr-package}
\alias{OmniCorr}
\alias{OmniCorr-package}
\title{OmniCorr: Multi-omics Integration and Visualization Framework}
\description{
OmniCorr is an R package-style framework for correlation-based integration and visualization of multi-omics datasets, including transcriptomics, metatranscriptomics, metagenomics, metaproteomics, and associated sample metadata. The package computes pairwise correlations between independently processed omics layers and provides flexible tools for organizing and visualizing cross-omics associations using aligned heatmaps and optional network-style visualizations. OmniCorr is designed to complement existing preprocessing and network inference workflows rather than replace them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shashank-KU/OmniCorr}
  \item Report bugs at \url{https://github.com/shashank-KU/OmniCorr/issues}
}

}
\author{
\strong{Maintainer}: Veronica Quarato \email{vequa7472@nmbu.no} (\href{https://orcid.org/0009-0008-4049-9746}{ORCID})

Authors:
\itemize{
  \item Veronica Quarato \email{vequa7472@nmbu.no} (\href{https://orcid.org/0009-0008-4049-9746}{ORCID})
  \item Shashank Gupta \email{bioinformaticssrm2011@gmail.com} (\href{https://orcid.org/0000-0001-8839-7003}{ORCID})
  \item Wanxin Lai
}

}
\keyword{internal}
