% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_correlations.R
\name{calculate_correlations}
\alias{calculate_correlations}
\title{Calculate correlation between two data frames}
\usage{
calculate_correlations(
  df1,
  df2,
  method = "pearson",
  adjust_method = "fdr",
  use = "all.obs",
  show_significance = TRUE
)
}
\arguments{
\item{df1}{A data frame containing the first set of variables to be correlated. The samples should be in the rows and the MEs in the columns.}

\item{df2}{A data frame containing the second set of variables to be correlated. The samples should be in the rows and the MEs in the columns.}

\item{method}{(optional) The correlation method to be used. Default is "pearson".}

\item{adjust_method}{(optional) The method for adjusting the p-values for multiple testing. Default is "fdr".}

\item{use}{(optional) A character string specifying the handling of missing data. Default is "all.obs".
The fast calculations currently support "all.obs" and "pairwise.complete.obs";
for other options, see R's standard correlation function cor. Abbreviations are allowed.}

\item{show_significance}{(optional) A logical indicating whether to show significance levels as stars or as p-value and correlation value. Default is TRUE (show stars).}
}
\value{
A list containing the correlation results

$correlation Matrix of correlation coefficients.

$p_value Matrix of p-values for the correlation coefficients.

$p_value_adj Matrix of adjusted p-values for the correlation coefficients.

$signif_matrix Matrix indicating the significance level of the correlations.
If show_significance is TRUE, the significance level is shown as stars;
otherwise, it is shown as the p-value and correlation value.
}
\description{
This function calculates the correlation matrix between the columns of two data frames.
}
\examples{
# Calculate correlation between two data frames using default parameters
result_list <- calculate_correlations(df1 = Transcriptomics, df2 = Metagenomics)

}
