% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_correlations.R
\name{calculate_correlations}
\alias{calculate_correlations}
\title{Calculate correlation between two data frames}
\usage{
calculate_correlations(
  df1,
  df2,
  method = "pearson",
  adjust_method = "fdr",
  use = "all.obs",
  show_significance = "stars"
)
}
\arguments{
\item{df1}{A data frame containing the first set of variables to be correlated. The samples should be in the rows and the Genes/Proteins/Metabolites/ASVs in the columns.}

\item{df2}{A data frame containing the second set of variables to be correlated. The samples should be in the rows and the Genes/Proteins/Metabolites/ASVs in the columns.}

\item{method}{(optional) The correlation method to be used. Default is "pearson".}

\item{adjust_method}{(optional) The method for adjusting the p-values for multiple testing. Default is "fdr".}

\item{use}{(optional) A character string specifying the handling of missing data. Default is "all.obs".
The fast calculations currently support "all.obs" and "pairwise.complete.obs";
for other options, see R's standard correlation function cor. Abbreviations are allowed.}

\item{show_significance}{(optional) A character string indicating whether to show significance levels as stars, p-values or correlation values. Default is "stars".
Possible values are "stars", "p_value" or "correlation".}
}
\value{
A list containing the correlation results

$correlation Matrix of correlation coefficients.

$p_value Matrix of p-values for the correlation coefficients.

$p_value_adj Matrix of adjusted p-values for the correlation coefficients.

$signif_matrix Matrix indicating the significance level of the correlations.
If show_significance is "stars", the significance level is shown as stars;
if show_significance is "p_value", the significance level is shown as p-value;
if show_significance is "correlation", the significance level is shown as the correlation value.
}
\description{
This function calculates the correlation matrix between the columns of two data frames.
}
\examples{
# Calculate correlation between two data frames using default parameters and show significance level as p-value
result_list <- calculate_correlations(df1 = Transcriptomics, df2 = Metagenomics, show_significance = "p_value")

}
